Port for dbd304c48100f04f7960517f03529eb99dd878c3  ([Java]: Reworked messageHeader in all schemas to reflect Issue #94)

+++ perf/csharp/Adaptive.SimpleBinaryEncoding.PerfTests/Bench/SBE/FIX/MassQuote.cs
@@ -10,8 +10,8 @@ namespace Adaptive.SimpleBinaryEncoding.PerfTests.Bench.SBE.FIX
     {
     public const ushort BlockLength = (ushort)62;
     public const ushort TemplateId = (ushort)105;
-    public const byte SchemaId = (byte)2;
-    public const byte Schema_Version = (byte)0;
+    public const ushort SchemaId = (ushort)2;
+    public const ushort Schema_Version = (ushort)0;
     public const string SematicType = "i";
 
     private readonly MassQuote _parentMessage;
